Refactored mechanism of isNew and isDependentFrom.
The concept of Registrar is introduced, whose aim is to avoid having to go through all the expression tree to know if the expression is new. Indeed, only few evaluables can change their values.
In particular, these are variables and mutables. The latter are similar to constants with the difference that they have defined the operator equal. All the methods isNew and isDependent from
have been substituted by the method addDependencies where the actual dependencies are specified.