Dependencies use raw pointers to Registrars. Added Assignable.
Having raw pointers allows to store (this) as dependency. In turn this allowed to define the Assignable object, where the operator = is defined, while avoiding to override the getDependencies method.
In turn, the static check for the = operator is easier. This seemed to be causing issues in VCC and clang.